Updates to readme and sample renaming
