+++ sdk/textanalytics/Azure.AI.TextAnalytics/tests/Azure.AI.TextAnalytics.Tests.csproj
@@ -4,6 +4,7 @@
   </PropertyGroup>
   <Import Project="..\..\..\core\Azure.Core\tests\TestFramework.props" />
   <ItemGroup>
+    <PackageReference Include="Azure.Identity" Version="1.1.0" />
     <PackageReference Include="nunit" />
     <PackageReference Include="NUnit3TestAdapter" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" />
