+++ sdk/textanalytics/Azure.AI.TextAnalytics/tests/samples/Sample3_ExtractKeyPhrasesBatchConvenience.cs
@@ -6,6 +6,7 @@ using NUnit.Framework;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Linq;
 
 namespace Azure.AI.TextAnalytics.Samples
 {
@@ -13,7 +14,7 @@ namespace Azure.AI.TextAnalytics.Samples
     public partial class TextAnalyticsSamples
     {
         [Test]
-        public void DetectLanguageBatch()
+        public void ExtractKeyPhrasesBatchConvenience()
         {
             string endpoint = Environment.GetEnvironmentVariable("TEXT_ANALYTICS_ENDPOINT");
             string subscriptionKey = Environment.GetEnvironmentVariable("TEXT_ANALYTICS_SUBSCRIPTION_KEY");
@@ -23,23 +24,24 @@ namespace Azure.AI.TextAnalytics.Samples
 
             var inputs = new List<string>
             {
-                "Hello world",
-                "Bonjour tout le monde",
-                "Hola mundo",
-                ":) :( :D",
+                "Microsoft was founded by Bill Gates and Paul Allen.",
+                "Text Analytics is one of the Azure Cognitive Services.",
+                "My cat might need to see a veterinarian.",
             };
 
-            Debug.WriteLine($"Detecting language for inputs:");
-            foreach (var input in inputs)
-            {
-                Debug.WriteLine($"    {input}");
-            }
-            DetectLanguageResultCollection results = client.DetectLanguages(inputs);
+            ExtractKeyPhrasesResultCollection results = client.ExtractKeyPhrases(inputs);
 
-            Debug.WriteLine($"Detected languages are:");
-            foreach (DetectLanguageResult result in results)
+            Debug.WriteLine($"Extracted key phrases for each input are:");
+            int i = 0;
+            foreach (var result in results)
             {
-                Debug.WriteLine($"    {result.PrimaryLanguage.Name}, with confidence {result.PrimaryLanguage.Score:0.00}.");
+                Debug.WriteLine($"For input: \"{inputs[i++]}\",");
+                Debug.WriteLine($"the following {result.KeyPhrases.Count()} key phrases were found: ");
+
+                foreach (var keyPhrase in result.KeyPhrases)
+                {
+                    Debug.WriteLine($"    {keyPhrase}");
+                }
             }
         }
     }
