+++ sdk/textanalytics/Azure.AI.TextAnalytics/tests/samples/Sample5_RecognizePiiEntitiesBatchConvenience.cs
@@ -14,7 +14,7 @@ namespace Azure.AI.TextAnalytics.Samples
     public partial class TextAnalyticsSamples
     {
         [Test]
-        public void RecognizeEntitiesBatch()
+        public void RecognizePiiEntitiesBatchConvenience()
         {
             string endpoint = Environment.GetEnvironmentVariable("TEXT_ANALYTICS_ENDPOINT");
             string subscriptionKey = Environment.GetEnvironmentVariable("TEXT_ANALYTICS_SUBSCRIPTION_KEY");
@@ -24,19 +24,18 @@ namespace Azure.AI.TextAnalytics.Samples
 
             var inputs = new List<string>
             {
-                "Microsoft was founded by Bill Gates and Paul Allen.",
-                "Text Analytics is one of the Azure Cognitive Services.",
-                "A key technology in Text Analytics is Named Entity Recognition (NER).",
+                "A developer with SSN 555-55-5555 whose phone number is 555-555-5555 is building tools with our APIs.",
+                "Your ABA number - 111000025 - is the first 9 digits in the lower left hand corner of your personal check.",
             };
 
-            RecognizeEntitiesResultCollection results = client.RecognizeEntities(inputs);
+            RecognizePiiEntitiesResultCollection results = client.RecognizePiiEntities(inputs);
 
-            Debug.WriteLine($"Recognized entities for each input are:");
+            Debug.WriteLine($"The following Personally Identifiable Information entities were recognized:");
             int i = 0;
             foreach (var result in results)
             {
                 Debug.WriteLine($"For input: \"{inputs[i++]}\",");
-                Debug.WriteLine($"the following {result.NamedEntities.Count()} entities were found: ");
+                Debug.WriteLine($"the following {result.NamedEntities.Count()} PII entit{(result.NamedEntities.Count() > 1 ? "ies were" : "y was")} found:");
 
                 foreach (var entity in result.NamedEntities)
                 {
